pyproject.toml
[project]
name = "edrsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logger>=1.4",
    "psutil>=7.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scapy>=2.7.0",
    "watchdog>=6.0.0",
]
